+++ sources/BUtil.Core/Options/ITaskShortcutService.cs
@@ -1,0 +1,7 @@
+namespace BUtil.Core.Options;
+
+public interface ITaskShortcutService
+{
+    void CreateOrUpdate(string taskName);
+    void Delete(string taskName);
+}