+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## New Features
 - Console: When application is launched without parameters, its possible to use execute tasks interactively.
+- UI: Shortcuts for tasks are created and maintained, so you can use start tasks for Linux and Windows via Start \ Search to launch them instead of application doing it directly.
 
 # 2026.05.21